% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_report.R
\name{automaticdatafConnect}
\alias{automaticdatafConnect}
\title{get data about a file on the server.}
\usage{
automaticdatafConnect(
  tablename,
  folder = getwd(),
  schema = NULL,
  dicoT = NULL,
  splitvar = NULL,
  Connect = NULL,
  Connectf = NULL,
  alwaysexclude = NULL
)
}
\value{
a list
}
\description{
get data about a file on the server.
}
