% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_report.R
\name{missing.summary}
\alias{missing.summary}
\title{Percentage of missing for each variable}
\usage{
missing.summary(X, info2 = NULL)
}
\arguments{
\item{X}{a data frame}

\item{info2}{a data frame with two variables named c("COLUMN_NAME","CONSTRAINT_TYPE")}
}
\value{
a data frame
}
\description{
Percentage of missing for each variable
}
\details{
Percentage of missing for each variable of a data frame.
}
\examples{
X=cars
for(i in 1:40){
  X[sample(1:50,1,replace=TRUE),sample(1:2,1,replace=TRUE)]<-NA}
missing.summary(X)
}
