% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_report.R
\name{var.summary}
\alias{var.summary}
\title{Summary for each variable in table.}
\usage{
var.summary(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a list
}
\description{
Summary for each variable in table.
}
\examples{
data(cars)
var.summary(cars)
}
