% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_report.R
\name{var.summaryConnect}
\alias{var.summaryConnect}
\title{Summary for each variable in table.}
\usage{
var.summaryConnect(X, datadic = NULL)
}
\arguments{
\item{X}{a data frame}

\item{datadic}{a data dictionnary}
}
\value{
a list
}
\description{
Summary for each variable in table.
}
\examples{
data(cars)
var.summaryConnect(cars)
}
