% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_report.R
\name{automaticRmd}
\alias{automaticRmd}
\title{Creates RMD files for all datasets in a specific folder.}
\usage{
automaticRmd(
  tablename,
  folder = getwd(),
  schema = NULL,
  specialprogram = lefichier(file.path(folder, paste0("study_", schema, "_", tablename,
    "_special.R"))),
  specialreport = lefichier(file.path(folder, paste0("study_", schema, "_", tablename,
    "_special.Rmd"))),
  specialdatafile = lefichier(file.path(folder, paste0("study_", schema, "_",
    tablename, "_special.rda"))),
  automaticdatafile = file.path(folder, paste0("study_", schema, "_", tablename,
    "_automatic.rda")),
  replace = TRUE,
  rerunspecial = FALSE,
  dicoT = NULL,
  splitvar = NULL,
  alwaysexclude = NULL,
  dico = function() {     NULL },
  author = ""
)
}
\value{

}
\description{
Creates RMD files for all datasets in a specific folder.
}
\examples{
tablename="cars"
automaticRmd(tablename)

}
